Associate sockets with a `Selector`
Closes #308

Windows requires that sockets are used exclusively with a single IOCP handle.
Unix platforms do not have this requirement. In order to provide a cross
platform API, Mio's network types artificially impose this limitations on Unix
platforms.

If a user of Mio would like to use a single socket with multiple event loops
on Unix platforms, this is possible by using `std::net` types and the
`EventedFd` wrapper.