Implement level triggering for windows
The windows selector uses a linked list of currently active level-triggered
events. When an event arrives and the socket is registered with
level-triggered, the event is tracked in this linked list. Every call to
select events will include all elements in this list. When an operation is
performed on the socket such that the socket is no longer ready, the event is
removed from the linked list.

Closes #241