Implement readv/writev support for TcpStream
This commit implements `readv` and `writev` support for TCP sockets in mio. This
corresponds to the `readv` and `writev` syscalls on Unix and a bit of custom
logic on Windows. On Windows writes are just concatenated into one buffer and
then handed off to IOCP, and each buffer to read into incurs a `recv` syscall
(currently). This may be optimizable in the future for Windows.

The other major factor in this commit is that a new `IoVec` type is added to
this library to represent a buffer passed to these functions. This type is a
dynamically sized type to allow `&IoVec` to represent `&[u8]` and `&mut IoVec`
to represent `&mut [u8]`, allowing the same type to be used in the read/write
versions of these functions. The in-memory representation of `IoVec` is
equivalent to `libc::iovec` on Unix and `winapi::WSABUF` on Windows, asserted
with various transmutes. This all allows us to build up 0-cost abstractions for
readv/writev.