Do not trigger HUP events on kqueue
Previously, mio would attempt to provide HUP events for kqueue platforms
that mimic the epoll behavior. However, it is not possible to do this
reliably. The assumption was that when an EVFILT_WRITE event is raised
with EOF set, this had to be the final event for a TCP socket due to how
the TCP protocol works.

As it turns out, user actions as well as external events can trigger
kqueue events. When the user calls `shutdown(SHUT_WR)`, this results in
an EVFILT_WRITE event with EOF set. Converting this to HUP at this point
is incorrect as data may still be received