Test cleanups (#1063)
* Move double register test into poll's test module

Reduces the output size.

* Replace custom port selection by using port 0

Port 0 is selection as it tells the OS to use any port available,
perfect for our use case.

* Move smoke tests into poll test module

* Make all Sync and Send assertions uniform

Moves them all to the top of the tests files and uses the same function
name. In addition this also checks if Waker is Send and Sync.

* Start each test with calling init

Now also doesn't ignore the possible error return by env_logger.