Add read_close and write_close readiness events (#1110)
## Motivation

The existing `is_hup` method on `Events` was difficult to add additional checks
to that were both cross-platform and guaranteed no false positives.

For example, if the write half of a socket is closed in certain scenarios, this
causes a set of expected events to be triggered. With `is_read_hup` and
`is_hup`, there was no way to check for for this set of events that did not also
result in false positives on other platforms.

## Solution

To solve this, `is_read_closed` and `is_write_closed` are being introduced as
cross-platform methods on `Events` that are best effort implementations--meaning
false negatives are possible, but false positives will not occur.

Finding the expected event behavior for each of these platforms has not been
very easy, so below is best-effort summary of what can be expected. #1091
provides a lot of the visibility into where this data came from, but a lot of
the scenarios checked for change when they occur independently.

### epoll

A helpful thing to keep in mind is that `EPOLLHUP` is a mask for a socket's
`RCV_SHUTDOWN` and `SEND_SHUTDOWN` flags. Once both of these flags are set,
`EPOLLHUP` will trigger.

#### is_read_closed

`EPOLLRDHUP` is now a registered event when there is interest in
`Interest::READABLE`.

`EPOLLRDHUP` occurs after *either* of the following events that sets the
`RCV_SHTUDOWN` flag:
  1. The local socket has called `shutdown(Shutdown::Read)`
  2. The peer stream has called `shutdown(Shutdown::Write)` which sends a `FIN`
     packet; that packet has been acknowledged by the local socket.

`EPOLLHUP` occurs after *both* of the following:
  1. `EPOLLRDHUP` by the previous point
  2. The local socket has called `shutdown(Shutdown::Write)` sending a `FIN` to
     the peer and marking setting its own `SEND_SHUTDOWN` flag.

`is_read_closed` checks for *either* of the following:
  1. `EPOLLHUP`
  2. `EPOLLIN` and `EPOLLRDHUP`

#### is_write_closed

`is_write_closed` is a hard event to check for with `epoll` without returning
false positives. The only special-case set of events that can be checked is both
`EPOLLOUT` and `EPOLLERR` which occurs when a Unix pipe sender closes.

Therefore, `is_write_closed` checks for *either* of the following:
  1. `EPOLLHUP` by the previous point
  2. `EPOLLOUT` and `EPOLLERR`

#### Things to note (or at least I found interesting)

A peer calling `shutdown(Shutdown::Both)` does not result in `EPOLLHUP` on the
local socket. It results in a `EPOLLIN` | `EPOLLOUT` | `EPOLLRDHUP` because the
peer has sent a `FIN` packet that has been acknowledged. Keep in mind however,
the local socket has not *also* sent a `FIN`, therefore it's `SEND_SHUTDOWN`
flag is not set.

Except for Unix pipes, a local socket calling `shutdown(Shutdown::Write)` only
results in a `EPOLLOUT`, which is not very helpful! Because there is no
`EPOLLRDHUP` equivalent event for when the `SEND_SHUTDOWN` flag is set, it is
hard to say whether or not the write half is _actually_ closed.

### kqueue

`kqueue` is has fairly predictable event behavior and therefore read and write
closed events are easier check. Note that some of the event checks in the tests
will only compile for `kqueue` platforms because `is_write_closed` will not
return false negatives.

#### is_read_closed

`is_read_closed` checks for the `EVFILT_READ` filter *and* `EV_EOF` flag

#### is_write_closed

`is_write_closed` checks for the `EVFILT_WRITE` filter *and* `EV_EOF` flag.

#### Things to note (or at least I found interesting)

A peer calling `shutdown(Shutdown::Both)` has the same result as it does in
`epoll`. `is_read_closed` will return `true`, but nothing has happend on the
local socket to indicate the write half will error.

### IOCP

`IOCP` is the least tested selector in this change, and therefore has few
situations where `is_read_closed` and `is_write_closed` do not return false
negatives. It checks for the equivalent behavior as before, mainly that
`is_write_closed` is equivalent to the current `is_hup`.

`is_read_closed` checks for the `AFD_POLL_DISCONNECT` flag.

`is_write_closed` checks for the `AFD_POLL_ABORT` flag.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>