Only start IOCTL_AFD_POLL op immediately if another thread is blocked (#1140)
Since it is not possible to modify or synchronously cancel an
IOCTL_AFD_POLL operation, and there can be only one active AFD_POLL
operation per (socket, completion port) pair at any time, it is
expensive to change a socket's event registration after it has been
submitted to the kernel.

Therefore, if no other threads are polling when interest in a socket
event is (re)registered, the socket is added to the 'update queue', but
the actual syscall to start the IOCTL_AFD_POLL operation is deferred
until just before the GetQueuedCompletionStatusEx() call is made.

However, when another thread is already blocked on
GetQueuedCompletionStatusEx() we tell the kernel about the registered
socket event(s) immediately.

Closes: #1128
Closes: #1135