Removed `try_clone` from `TcpListener`, `TcpStream`, `UdpSocket` (associated tests were also removed) This api had several limitations . More details in issue #1074.
On Unix, `try_clone` was just a dup, so the behavior depended on the underlying
epoll and kqueue implementations. This meant there was no defined behavior
because of the platform dependent differences between epoll and kqueue
implementations.

On Windows, polling for the readiness events of a cloned `TcpListener` was
broken. A cloned `TcpListener` polled events only for the i/o operations
performed on itself. To retrieve the same events as polling the initial object
a clone would have needed to be reregistered whenever the initial object was
reregistered (explicit reregister or wouldblock reregister).

Signed-off-by: Daniel Tacalau <dst4096@gmail.com>