illumos epoll may not wake up unless pipe is empty
At present, mio makes use of the emulated epoll(5) available on illumos
systems.  On some versions of illumos the interaction between pipes and
EPOLLET is not quite right.  Specifically, a wakeup may only be
triggered when the buffer pipe transitions from empty to non-empty.  To
work around this, flush the buffer for every call to wake().