ignore tcp_shutdown_client_both_close_event failure on illumos
At present, is_write_closed() does not appear to be true on the event
that fires after shutdown() is called.  There is already a similar gag
for the tcp_shutdown_client_write_close_event test, so add one here for
now.  When support for the native illumos "event ports" polling
mechanism is added, to replace the use of epoll emulation, this can be
revisted.